routes refactoring. Postman collection updated
