+++ web-api/postman-tests/Golang & Gin.postman_collection.json
@@ -1,10 +1,9 @@
 {
 	"info": {
-		"_postman_id": "4933c26d-4b9e-43a4-ac89-5c67e877a703",
+		"_postman_id": "ebd80c85-5f9e-4762-bb6b-c7a7c57fae68",
 		"name": "Golang & Gin",
 		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "39808379",
-		"_collection_link": "https://hoda-team.postman.co/workspace/HODA-Team-Workspace~fa1861b7-b8b0-4d50-83d8-f91b24d6d6c1/collection/39808379-4933c26d-4b9e-43a4-ac89-5c67e877a703?action=share&source=collection_link&creator=39808379"
+		"_exporter_id": "3493272"
 	},
 	"item": [
 		{
@@ -59,6 +58,31 @@
 					"port": "8080",
 					"path": [
 						"events"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Get Event By Id",
+			"request": {
+				"method": "GET",
+				"header": [
+					{
+						"key": "Content-Type",
+						"value": "application/json",
+						"type": "text"
+					}
+				],
+				"url": {
+					"raw": "localhost:8080/events/3",
+					"host": [
+						"localhost"
+					],
+					"port": "8080",
+					"path": [
+						"events",
+						"3"
 					]
 				}
 			},
