+++ web-api/routes/routes.go
@@ -1,0 +1,10 @@
+package routes
+
+import "github.com/gin-gonic/gin"
+
+func RegisterRoutes(server *gin.Engine) {
+	// handlers registration
+	server.GET("/events", getEvents)
+	server.GET("/events/:id", getEventById)
+	server.POST("/events", createEvent)
+}
